backend/requirements.txt
Flask>=3.0,<4
Flask-SQLAlchemy>=3.1,<4
Flask-JWT-Extended>=4.6,<5
Flask-Cors>=4.0,<5
PyMySQL>=1.1,<2
python-dotenv>=1.0,<2

